+++ app/app/src/main/java/com/winlator/winhandler/TaskManagerDialog.java
@@ -190,22 +190,23 @@ public class TaskManagerDialog extends ContentDialog implements OnGetProcessInfo
         listItemMenu.setOnMenuItemClickListener((menuItem) -> {
             int itemId = menuItem.getItemId();
             final WinHandler winHandler = activity.getWinHandler();
-            switch (itemId) {
-                case R.id.menu_item_process_affinity:
+            int _id = itemId;
+            if (_id == R.id.menu_item_process_affinity) {
                     showProcessorAffinityDialog(processInfo);
-                    break;
-                case R.id.menu_item_bring_to_front:
+            }
+            else if (_id == R.id.menu_item_bring_to_front) {
                     winHandler.bringToFront(processInfo.name);
                     dismiss();
-                    break;
-                case R.id.menu_item_end_process:
+            }
+            else if (_id == R.id.menu_item_end_process) {
                     ContentDialog.confirm(activity, R.string.do_you_want_to_end_this_process, () -> {
                         winHandler.killProcess(null, processInfo.pid);
                     });
-                    break;
             }
+
             return true;
         });
+        listItemMenu.show();
         listItemMenu.show();
     }
